1709: Fixed unit tests
